@@ -101,13 +101,13 @@ class OneDungeonAudioPlayer {
 
   bool isSfxActive;
 
-  PlayBackgroundAudio _playBackgroundAudio;
+  final PlayBackgroundAudio _playBackgroundAudio;
 
-  StopBackgroundAudio _stopBackgroundAudio;
+  final StopBackgroundAudio _stopBackgroundAudio;
 
-  PlaySingleAudio _playSingleAudio;
+  final PlaySingleAudio _playSingleAudio;
 
-  PreCacheSingleAudio preCacheSingleAudio;
+  final PreCacheSingleAudio preCacheSingleAudio;
 
   /// Plays the received audio.
   Future<void> play(OneDungeonAudio audio) async {
@@ -116,36 +116,28 @@ class OneDungeonAudioPlayer {
     switch (audio) {
       case OneDungeonAudio.backgroundMusic:
         if (!isBackgroundMusicActive) return;
-        await _playBackgroundAudio(GameAssets.kBackgroudMusic, volume: 0.5);
-        break;
+        return _playBackgroundAudio(GameAssets.kBackgroudMusic, volume: 0.5);
       case OneDungeonAudio.descending:
         if (!isSfxActive) return;
-        await _playSingleAudio(GameAssets.kDescendingSfx);
-        break;
+        return _playSingleAudio(GameAssets.kDescendingSfx);
       case OneDungeonAudio.jump:
         if (!isSfxActive) return;
-        await _playSingleAudio(GameAssets.kJumpSfx);
-        break;
+        return _playSingleAudio(GameAssets.kJumpSfx);
       case OneDungeonAudio.landing:
         if (!isSfxActive) return;
-        await _playSingleAudio(GameAssets.kLandingSfx);
-        break;
+        return _playSingleAudio(GameAssets.kLandingSfx);
       case OneDungeonAudio.pickUp:
         if (!isSfxActive) return;
-        await _playSingleAudio(GameAssets.kPickUpSfx);
-        break;
+        return _playSingleAudio(GameAssets.kPickUpSfx);
       case OneDungeonAudio.pickUpStar:
         if (!isSfxActive) return;
-        await _playSingleAudio(GameAssets.kPickUpStarSfx);
-        break;
+        return _playSingleAudio(GameAssets.kPickUpStarSfx);
       case OneDungeonAudio.select:
         if (!isSfxActive) return;
-        await _playSingleAudio(GameAssets.kSelectSfx);
-        break;
+        return _playSingleAudio(GameAssets.kSelectSfx);
       case OneDungeonAudio.success:
         if (!isSfxActive) return;
-        await _playSingleAudio(GameAssets.kSuccessSfx);
-        break;
+        return _playSingleAudio(GameAssets.kSuccessSfx);
     }
   }
 
