@@ -5,11 +5,12 @@ import 'package:flame_test/flame_test.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:one_dungeon/entities/entities.dart';
 import 'package:one_dungeon/injector.dart' as di;
+import 'package:one_dungeon/one_dungeon_audio/one_dungeon_audio.dart';
 
-import '../../../helpers/test_game.dart';
+import '../../../helpers/helpers.dart';
 
 void main() {
-  TestWidgetsFlutterBinding.ensureInitialized();
+  TestWidgetsBinding.ensureInitialized();
 
   late BoyAirResistanceBehavior boyAirResistanceBehavior;
 
@@ -19,6 +20,8 @@ void main() {
 
   setUpAll(() async {
     await di.initializeDependencies();
+    await di.injector.unregister<OneDungeonAudioPlayer>();
+    di.injector.registerSingleton<OneDungeonAudioPlayer>(TestAudioPlayer());
   });
 
   tearDownAll(() async {
@@ -36,7 +39,8 @@ void main() {
         );
 
         await game.ready();
-        await game.ensureAdd(boy);
+
+        await game.world.ensureAdd(boy);
 
         expect(boy.velocity.x, equals(10));
 
@@ -56,7 +60,7 @@ void main() {
         );
 
         await game.ready();
-        await game.ensureAdd(boy);
+        await game.world.ensureAdd(boy);
 
         expect(boy.velocity.x, equals(-10));
 
@@ -77,7 +81,7 @@ void main() {
         );
 
         await game.ready();
-        await game.ensureAdd(boy);
+        await game.world.ensureAdd(boy);
 
         expect(boy.position.x, equals(10));
 
@@ -93,18 +97,27 @@ void main() {
       (game) async {
         final boy = Boy.test(
           velocity: Vector2(10, 0),
-          center: Vector2(game.size[0], 0),
+          center: Vector2(game.cameraComponent.visibleWorldRect.width, 0),
           behavior: boyAirResistanceBehavior,
         );
 
         await game.ready();
-        await game.ensureAdd(boy);
 
-        expect(boy.position.x, equals(game.size[0]));
+        await game.world.ensureAdd(boy);
+
+        expect(
+          boy.position.x,
+          equals(game.cameraComponent.visibleWorldRect.width),
+        );
 
         game.update(1);
 
-        expect(boy.position.x, equals(game.size[0] - (boy.size.x / 2)));
+        expect(
+          boy.position.x,
+          equals(
+            game.cameraComponent.visibleWorldRect.width - (boy.size.x / 2),
+          ),
+        );
       },
     );
   });
