@@ -12,9 +12,8 @@ import 'package:one_dungeon/one_dungeon_audio/one_dungeon_audio.dart';
 import '../../../helpers/helpers.dart';
 
 void main() {
-  TestWidgetsFlutterBinding.ensureInitialized();
+  TestWidgetsBinding.ensureInitialized();
 
-  late OneDungeonAudioPlayer audioPlayer;
   late StarCollidingBehavior starCollidingBehavior;
 
   setUp(() {
@@ -22,18 +21,9 @@ void main() {
   });
 
   setUpAll(() async {
-    audioPlayer = OneDungeonAudioPlayer.test(
-      playBackgroundAudio: (_, {double? volume}) async {},
-      stopBackgroundAudio: () async {},
-      playSingleAudio: (_, {double? volume}) async {},
-      preCacheSingleAudio: (_) async {},
-    );
-
     await di.initializeDependencies();
-
     await di.injector.unregister<OneDungeonAudioPlayer>();
-
-    di.injector.registerSingleton<OneDungeonAudioPlayer>(audioPlayer);
+    di.injector.registerSingleton<OneDungeonAudioPlayer>(TestAudioPlayer());
   });
 
   tearDownAll(() async {
@@ -57,8 +47,9 @@ void main() {
         );
 
         await game.ready();
-        await game.ensureAdd(star);
-        await game.ensureAdd(boy);
+
+        await game.world.ensureAdd(star);
+        await game.world.ensureAdd(boy);
 
         final collectedStars = game.collectedStars;
 
@@ -85,16 +76,17 @@ void main() {
         );
 
         await game.ready();
-        await game.ensureAdd(star);
-        await game.ensureAdd(boy);
 
-        expect(game.children.contains(star), isTrue);
+        await game.world.ensureAdd(star);
+        await game.world.ensureAdd(boy);
+
+        expect(game.world.children.contains(star), isTrue);
 
         await starCollidingBehavior
             .onCollisionStart({Vector2(0, 0), Vector2(10, 0)}, star);
         game.update(1);
 
-        expect(game.children.contains(star), isFalse);
+        expect(game.world.children.contains(star), isFalse);
       },
     );
   });
