@@ -5,11 +5,12 @@ import 'package:flame_test/flame_test.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:one_dungeon/entities/entities.dart';
 import 'package:one_dungeon/injector.dart' as di;
+import 'package:one_dungeon/one_dungeon_audio/one_dungeon_audio.dart';
 
 import '../../../helpers/helpers.dart';
 
 void main() {
-  TestWidgetsFlutterBinding.ensureInitialized();
+  TestWidgetsBinding.ensureInitialized();
 
   late ElevatorMovingBehavior elevatorMovingBehavior;
 
@@ -19,6 +20,8 @@ void main() {
 
   setUpAll(() async {
     await di.initializeDependencies();
+    await di.injector.unregister<OneDungeonAudioPlayer>();
+    di.injector.registerSingleton<OneDungeonAudioPlayer>(TestAudioPlayer());
   });
 
   tearDownAll(() async {
@@ -37,7 +40,8 @@ void main() {
         );
 
         await game.ready();
-        await game.ensureAdd(elevator);
+
+        await game.world.ensureAdd(elevator);
 
         elevatorMovingBehavior.isGoingDown = true;
 
@@ -58,7 +62,8 @@ void main() {
         );
 
         await game.ready();
-        await game.ensureAdd(elevator);
+
+        await game.world.ensureAdd(elevator);
 
         elevatorMovingBehavior.isGoingDown = false;
 
